refactor: rename old ControlsDropdown component
